customer-service/requirements.txt
fastapi
uvicorn[standard]
tortoise-orm[asyncpg]
aerich